chore: make UnmarshalPacketData public
